--- src/teleoperation/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/teleoperation
-[install]
-install_scripts=$base/lib/teleoperation
